Refactor: Updates after PR comments
