+++ packages/grafana-ui/src/components/ErrorBoundary/ErrorWithStack.tsx
@@ -1,13 +1,21 @@
 import React, { FunctionComponent } from 'react';
 import { ErrorBoundaryApi } from './ErrorBoundary';
+import { stylesFactory } from '../../themes';
+import { css } from 'emotion';
+
+const getStyles = stylesFactory(() => {
+  return css`
+    width: 500px;
+    margin: 64px auto;
+  `;
+});
 
 export interface Props extends ErrorBoundaryApi {
-  className: string;
   title: string;
 }
 
-export const ErrorWithStack: FunctionComponent<Props> = ({ error, errorInfo, className, title }) => (
-  <div className={className}>
+export const ErrorWithStack: FunctionComponent<Props> = ({ error, errorInfo, title }) => (
+  <div className={getStyles()}>
     <h2>{title}</h2>
     <details style={{ whiteSpace: 'pre-wrap' }}>
       {error && error.toString()}
