+++ public/app/core/components/DynamicImports/ErrorLoadingChunk.tsx
@@ -1,19 +1,26 @@
 import React, { FunctionComponent } from 'react';
-import { Button } from '@grafana/ui';
+import { Button, stylesFactory } from '@grafana/ui';
+import { css } from 'emotion';
+
+const getStyles = stylesFactory(() => {
+  return css`
+    width: 508px;
+    margin: 128px auto;
+  `;
+});
 
 interface Props {
   error: Error | null;
-  className: string;
 }
 
-export const ErrorLoadingChunk: FunctionComponent<Props> = ({ error, className }) => (
-  <div className={className}>
+export const ErrorLoadingChunk: FunctionComponent<Props> = ({ error }) => (
+  <div className={getStyles()}>
     <h2>Unable to find application file</h2>
     <br />
     <h2 className="page-heading">Grafana has likely been updated. Please try reloading the page.</h2>
     <br />
     <div className="gf-form-group">
-      <Button size={'md'} variant={'secondary'} icon="fa fa-repeat" onClick={() => window.location.reload()}>
+      <Button size="md" variant="secondary" icon="fa fa-repeat" onClick={() => window.location.reload()}>
         Reload
       </Button>
     </div>
