+++ public/app/core/components/DynamicImports/LoadingChunkPlaceHolder.tsx
@@ -1,9 +1,8 @@
 import React, { FunctionComponent } from 'react';
-import { cx } from 'emotion';
 import { LoadingPlaceholder } from '@grafana/ui';
 
 export const LoadingChunkPlaceHolder: FunctionComponent = React.memo(() => (
-  <div className={cx('preloader')}>
+  <div className="preloader">
     <LoadingPlaceholder text={'Loading...'} />
   </div>
 ));
